+++ tests/widgets.py
@@ -1,6 +1,7 @@
 from __future__ import unicode_literals
 
 from unittest import TestCase
+from markupsafe import Markup
 from wtforms.widgets import html_params, Input
 from wtforms.widgets import *
 from wtforms.widgets import html5
@@ -22,14 +23,6 @@ class DummyField(object):
     iter_choices = lambda x: iter(x.data)
 
 
-class EscapeHtmlTest(TestCase):
-    def test(self):
-        self.assertEqual(core.escape_html('<i class="bar">foo</i>'), '&lt;i class=&quot;bar&quot;&gt;foo&lt;/i&gt;')
-        self.assertEqual(core.escape_html('<i class="bar">foo</i>', quote=False), '&lt;i class="bar"&gt;foo&lt;/i&gt;')
-        self.assertEqual(core.escape_html(HTMLString('<i class="bar">foo</i>')), '<i class="bar">foo</i>')
-        self.assertEqual(core.escape_html(HTMLString('<i class="bar">foo</i>'), quote=False), '<i class="bar">foo</i>')
-
-
 class HTMLParamsTest(TestCase):
     def test_basic(self):
         self.assertEqual(html_params(foo=9, k='wuuu'), 'foo="9" k="wuuu"')
@@ -42,8 +35,12 @@ class HTMLParamsTest(TestCase):
         self.assertEqual(html_params(data_foo=22), 'data-foo="22"')
         self.assertEqual(html_params(data_foo_bar=1), 'data-foo-bar="1"')
 
+    def test_aria_prefix(self):
+        self.assertEqual(html_params(aria_foo='bar'), 'aria-foo="bar"')
+        self.assertEqual(html_params(aria_foo_bar='foobar'), 'aria-foo-bar="foobar"')
+
     def test_quoting(self):
-        self.assertEqual(html_params(foo='hi&bye"quot'), 'foo="hi&amp;bye&quot;quot"')
+        self.assertEqual(html_params(foo='hi&bye"quot'), 'foo="hi&amp;bye&#34;quot"')
 
 
 class ListWidgetTest(TestCase):
@@ -117,7 +114,7 @@ class BasicWidgetsTest(TestCase):
     def test_textarea(self):
         # Make sure textareas escape properly and render properly
         f = DummyField('hi<>bye')
-        self.assertEqual(TextArea()(f), '<textarea id="" name="f">hi&lt;&gt;bye</textarea>')
+        self.assertEqual(TextArea()(f), '<textarea id="" name="f">\r\nhi&lt;&gt;bye</textarea>')
 
     def test_file(self):
         self.assertEqual(FileInput()(self.field), '<input id="id" name="bar" type="file">')
@@ -149,10 +146,10 @@ class SelectTest(TestCase):
         )
         self.assertEqual(
             Select.render_option('bar', '<i class="bar"></i>foo', False),
-            '<option value="bar">&lt;i class="bar"&gt;&lt;/i&gt;foo</option>'
+            '<option value="bar">&lt;i class=&#34;bar&#34;&gt;&lt;/i&gt;foo</option>'
         )
         self.assertEqual(
-            Select.render_option('bar', HTMLString('<i class="bar"></i>foo'), False),
+            Select.render_option('bar', Markup('<i class="bar"></i>foo'), False),
             '<option value="bar"><i class="bar"></i>foo</option>'
         )